Extract tests in their own GH step
