+++ docker-compose.dev.yml
@@ -8,6 +8,7 @@ services:
       NODE_ENV: development
       DATABASE_LOGGING: "${DATABASE_LOGGING-True}"
       JWT_SECRET: BOOKKEEPING-DEV
+      PAGE_ITEMS_LIMIT: 100
       GRPC_INTERNAL_ORIGIN: '[::]:4001'
       GRPC_AUTHENTICATED_ORIGIN: '[::]:4002'
     ports:
