--- test/public/runs/dataPassesUtilities.js
@@ -1,21 +1,0 @@
-const { waitForNavigation, pressElement, getInnerText, expectUrlParams, waitForTableLength } = require('../defaults.js');
-
-/**
- * Navigate to Runs per Data Pass page
- *
- * @param {Puppeteer.Page} page page
- * @param {number} lhcPeriodId id of lhc period on LHC Period overview page
- * @param {number} dataPassId id of data pass on Data Passes per LHC Period page
- * @param {number} expectedRowsCount expected number of rows on runs per data pass page
- * @return {Promise<void>} promise
- */
-exports.navigateToRunsPerDataPass = async (page, lhcPeriodId, dataPassId, expectedRowsCount) => {
-    await waitForNavigation(page, () => pressElement(page, 'a#lhc-period-overview', true));
-    const pdpBeamTypes = await getInnerText(await page.waitForSelector(`#row${lhcPeriodId}-pdpBeamTypes`));
-    await waitForNavigation(page, () => pressElement(page, `#row${lhcPeriodId}-associatedDataPasses a`, true));
-    expectUrlParams(page, { page: 'data-passes-per-lhc-period-overview', lhcPeriodId });
-    await page.waitForSelector('th#description');
-    await waitForNavigation(page, () => pressElement(page, `#row${dataPassId}-associatedRuns a`, true));
-    expectUrlParams(page, { page: 'runs-per-data-pass', dataPassId, pdpBeamTypes });
-    await waitForTableLength(page, expectedRowsCount);
-};
