+++ test/public/runsPerPeriodOrPass/index.js
@@ -11,10 +11,12 @@
  * or submit itself to any jurisdiction.
  */
 
-const OverviewSuite = require('./overview.test');
-const DetailSuite = require('./detail.test');
+const RunsPerPeriodOverviewSuite = require('./runsPerLhcPeriod.overview.test');
+const RunsPerDataPassOverviewPage = require('./runsPerDataPass.overview.test');
+const RunsPerSimulationPassOverviewPage = require('./runsPerSimulationPass.overview.test');
 
 module.exports = () => {
-    describe('Overview Page', OverviewSuite);
-    describe('Detail Page', DetailSuite);
+    describe('Runs Per LHC Period Overview Page', RunsPerPeriodOverviewSuite);
+    describe('Runs Per Data Pass Overview Page', RunsPerDataPassOverviewPage);
+    describe('Runs Per Simulation Pass Overview Page', RunsPerSimulationPassOverviewPage);
 };
