+++ test/scripts/test-runsPerPeriodOrPass.js
@@ -12,8 +12,9 @@
  */
 
 const application = require('../../lib/application');
-const OverviewSuite = require('../public/runs/overview.test');
-const DetailSuite = require('../public/runs/detail.test');
+const RunsPerPeriodOverviewSuite = require('../public/runs/runsPerLhcPeriod.overview.test');
+const RunsPerDataPassOverviewPage = require('../public/runs/runsPerDataPass.overview.test');
+const RunsPerSimulationPassOverviewPage = require('../public/runs/runsPerSimulationPass.overview.test');
 
 describe('Bookkeeping', () => {
     before(async () => {
@@ -26,7 +27,8 @@ describe('Bookkeeping', () => {
     });
 
     describe('Runs', () => {
-        describe('Overview Page', OverviewSuite);
-        describe('Detail Page', DetailSuite);
+        describe('Runs Per LHC Period Overview Page', RunsPerPeriodOverviewSuite);
+        describe('Runs Per Data Pass Overview Page', RunsPerDataPassOverviewPage);
+        describe('Runs Per Simulation Pass Overview Page', RunsPerSimulationPassOverviewPage);
     });
 });
